[clang][HeaderSearch] Treat permission-denied include paths as non-fatal
If part of the -I search path cannot be accessed (permissions),
clang will throw an error, even if the header can be found in
subsequent search path elements. This is counter productive as
the compilation will seem to have failed due to that error, even
though everyone is happy.

Based on the original patch by Axel Naumann