[clang][HeaderSearch] Treat permission-denied include paths as non-fatal
If part of the -I search path cannot be accessed due to permissions,
Clang no longer treats this as a fatal error. Instead, it emits a warning
pointing to the exact path that cannot be accessed while continuing the
search in subsequent search path elements.

Based on the original patch by Axel Naumann.